Issue #54 fixed: Moved example tests (some of them fail intentionally) to another project.
