+++ .gitignore
@@ -1,7 +1,7 @@
 bin
-/build
+build
 .settings
 test_output
-/target
-/.project
-/.classpath
+target
+.project
+.classpath
