+++ jbehave-junit-runner-examples/pom.xml
@@ -1,0 +1,77 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>de.codecentric</groupId>
+	<artifactId>jbehave-junit-runner-examples</artifactId>
+	<version>1.2.0-SNAPSHOT</version>
+	<packaging>jar</packaging>
+
+	<name>JBehave JUnit Runner</name>
+	<description/>
+	<url>https://github.com/codecentric/jbehave-junit-runner</url>
+	<organization>
+		<name>codecentric AG</name>
+		<url>http://www.codecentric.de</url>
+	</organization>
+	<licenses>
+		<license>
+			<name>MIT License</name>
+			<url>http://www.opensource.org/licenses/MIT</url>
+			<distribution>repo</distribution>
+		</license>
+	</licenses>
+	<developers>
+		<developer>
+			<name>Andreas Ebbert-Karroum</name>
+			<email>andreas.ebbert-karroum@codecentric.de</email>
+			<organization>codecentric AG</organization>
+		</developer>
+		<developer>
+			<name>Daniel Schneller</name>
+			<email>daniel.schneller@codecentric.de</email>
+			<organization>codecentric AG</organization>
+		</developer>
+	</developers>
+	<contributors>
+		<contributor>
+			<name>Stephen Kurlow</name>
+			<email>skurlow@gmail.com</email>
+		</contributor>
+	</contributors>
+
+	<scm>
+		<connection>scm:git:git://github.com/codecentric/jbehave-junit-runner.git</connection>
+		<developerConnection>scm:git:ssh://git@github.com/codecentric/jbehave-junit-runner.git</developerConnection>
+		<url>https://github.com/codecentric/jbehave-junit-runner</url>
+		<tag>jbehave-junit-runner-1.0.0</tag>
+	</scm>
+
+	<properties>
+		<jbehave.version>3.9</jbehave.version>
+		<junit.version>4.11</junit.version>
+		<mockito.version>1.9.0</mockito.version>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+	</properties>
+
+	<dependencies>
+		<dependency>
+			<groupId>de.codecentric</groupId>
+			<artifactId>jbehave-junit-runner</artifactId>
+			<version>1.2.0-SNAPSHOT</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.jbehave</groupId>
+			<artifactId>jbehave-core</artifactId>
+			<scope>test</scope>
+			<version>${jbehave.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<version>${junit.version}</version>
+		</dependency>
+	</dependencies>
+
+</project>
