+++ jbehave-junit-runner-examples/src/test/java/de/codecentric/jbehave/junit/monitoring/ExampleScenarioJUnitStories.java
@@ -1,8 +1,6 @@
 package de.codecentric.jbehave.junit.monitoring;
 
-import java.util.Arrays;
-import java.util.List;
-
+import de.codecentric.jbehave.junit.monitoring.step.ExampleSteps;
 import org.jbehave.core.configuration.Configuration;
 import org.jbehave.core.configuration.MostUsefulConfiguration;
 import org.jbehave.core.junit.JUnitStories;
@@ -10,7 +8,8 @@ import org.jbehave.core.steps.InjectableStepsFactory;
 import org.jbehave.core.steps.InstanceStepsFactory;
 import org.junit.runner.RunWith;
 
-import de.codecentric.jbehave.junit.monitoring.step.ExampleSteps;
+import java.util.Arrays;
+import java.util.List;
 
 @RunWith(JUnitReportingRunner.class)
 public class ExampleScenarioJUnitStories extends JUnitStories {
