+++ jbehave-junit-runner-examples/src/test/java/de/codecentric/jbehave/junit/monitoring/ExampleScenarioJUnitStoriesLocalized.java
