+++ jbehave-junit-runner-examples/src/test/resources/de/codecentric/jbehave/junit/monitoring/Multiplication_de.story
