+++ jbehave-junit-runner-examples/src/test/resources/de/codecentric/jbehave/junit/monitoring/failed_step_in_first_scenario_and_second_scenario_passes_j_unit_story.story
