+++ jbehave-junit-runner-examples/src/test/resources/de/codecentric/jbehave/junit/monitoring/failed_step_in_scenario_j_unit_story.story
