+++ jbehave-junit-runner-examples/src/test/resources/de/codecentric/jbehave/junit/monitoring/show_ignored_scenarios_as_ignored_and_not_passed_story.story
