+++ jbehave-junit-runner/src/test/java/de/codecentric/jbehave/junit/monitoring/suite/JUnitReportingRunnerIntegrationTest.java
